WIP
- some basic content validation tests added and passing
- verify output base contains some expected content

+++ rules/rbe_repo/container.bzl
@@ -24,14 +24,39 @@ load(
 # https://github.com/bazelbuild/bazel/issues/1262
 _EXTERNAL_FOLDER_PREFIX = "external/"
 
-_ROOT_DIR = "/rbe_autoconf"
-_PROJECT_REPO_DIR = "project_src"
-_REPO_DIR = _ROOT_DIR + "/" + _PROJECT_REPO_DIR
-_OUTPUT_DIR = _ROOT_DIR + "/autoconf_out"
-
-_BAZELISK_PATH = _ROOT_DIR + "/bazelisk"
 _BAZELISK_RELEASE = "v1.3.0"
-_BAZELISK_SHA = "98af93c6781156ff3dd36fa06ba6b6c0a529595abb02c569c99763203f3964cc"
+_BAZELISK_FILENAME_LINUX = "bazelisk"
+_BAZELISK_SHA_LINUX = "98af93c6781156ff3dd36fa06ba6b6c0a529595abb02c569c99763203f3964cc"
+_BAZELISK_FILENAME_WINDOWS = "bazelisk.exe"
+_BAZELISK_SHA_WINDOWS = "31fa9fcf250fe64aa3c5c83b69d76e1e9571b316a58bb5c714084495623e38b0"
+_PROJECT_REPO_DIR = "project_src"
+
+def _is_windows(ctx):
+    return ctx.os.name.lower().find("windows") != -1
+
+def _root_dir(ctx):
+    if _is_windows(ctx):
+        return "C:/rbe_autoconf"
+    else:
+        return "/rbe_autoconf"
+
+def _download_bazelisk(ctx):
+    bazelisk_url = "https://github.com/bazelbuild/bazelisk/releases/download/%s/bazelisk-" % _BAZELISK_RELEASE
+    if _is_windows(ctx):
+        bazelisk_url += "windows-amd64.exe"
+        bazelisk_filename = "bazelisk" + "/" + _BAZELISK_FILENAME_WINDOWS
+        bazelisk_sha = _BAZELISK_SHA_WINDOWS
+    else:
+        bazelisk_url += "linux-amd64"
+        bazelisk_filename = _BAZELISK_FILENAME_WINDOWS
+        bazelisk_sha = _BAZELISK_SHA_WINDOWS
+    ctx.download(bazelisk_url, bazelisk_filename, bazelisk_sha)
+
+def _bazelisk_path(ctx):
+    if _is_windows(ctx):
+        return _root_dir(ctx) + "/bazelisk/" + _BAZELISK_FILENAME_WINDOWS
+    else:
+        return _root_dir(ctx) + "/" + _BAZELISK_FILENAME_WINDOWS
 
 # Creates file "container/run_in_container.sh" which will be copied onto container
 # to run the commands to run bazel and create the output tar
@@ -40,15 +65,19 @@ def _create_docker_cmd(
         config_repos,
         outputs_tar,
         use_default_project):
-    # Set permissions on bazelisk
-    bazelisk_cmd = "chmod +x " + _BAZELISK_PATH
+    bazelisk_cmd = ""
+
+    if not _is_windows(ctx):
+        # Set permissions on bazelisk
+        bazelisk_cmd = "chmod +x " + _bazelisk_path(ctx)
 
     # Command to recursively convert soft links to hard links in the config_repos
     # Needed because some outputs of local_cc_config (e.g., dummy_toolchain.bzl)
     # could be symlinks.
     deref_symlinks_cmd = []
     for config_repo in config_repos:
-        symlinks_cmd = ("find $(" + _BAZELISK_PATH + " info output_base)/" +
+        # TODO: fix the find command below
+        symlinks_cmd = ("/usr/bin/find $(" + _bazelisk_path(ctx) + " info output_base)/" +
                         _EXTERNAL_FOLDER_PREFIX + config_repo +
                         " -type l -exec bash -c 'ln -f \"$(readlink -m \"$0\")\" \"$0\"' {} \\;")
         deref_symlinks_cmd.append(symlinks_cmd)
@@ -56,11 +85,12 @@ def _create_docker_cmd(
 
     # Command to copy produced toolchain configs to a tar at the root
     # of the container.
-    copy_cmd = ["mkdir " + _OUTPUT_DIR]
+    output_dir = _root_dir(ctx) + "/autoconf_out"
+    copy_cmd = ["mkdir " + output_dir]
     for config_repo in config_repos:
-        src_dir = "$(" + _BAZELISK_PATH + " info output_base)/" + _EXTERNAL_FOLDER_PREFIX + config_repo
-        copy_cmd.append("cp -dr " + src_dir + " " + _OUTPUT_DIR)
-    copy_cmd.append("tar -cf /" + outputs_tar + " -C " + _OUTPUT_DIR + "/ . ")
+        src_dir = "$(" + _bazelisk_path(ctx) + " info output_base)/" + _EXTERNAL_FOLDER_PREFIX + config_repo
+        copy_cmd.append("cp -dr " + src_dir + " " + output_dir)
+    copy_cmd.append("tar -cf /" + outputs_tar + " -C " + output_dir + "/ . ")
     output_copy_cmd = " && ".join(copy_cmd)
 
     # A success command to run after the output_copy_cmd finished.
@@ -69,23 +99,23 @@ def _create_docker_cmd(
 
     # if use_default_project was selected, we need to modify the WORKSPACE and BUILD file
     setup_default_project_cmd = ["cd ."]
-    if use_default_project:
-        setup_default_project_cmd += ["cd " + _ROOT_DIR + "/" + _PROJECT_REPO_DIR]
-        setup_default_project_cmd += ["mv BUILD.sample BUILD"]
-        setup_default_project_cmd += ["touch WORKSPACE"]
 
-    bazel_cmd = "cd " + _ROOT_DIR + "/" + _PROJECT_REPO_DIR
+    bazel_cmd = "cd " + _root_dir(ctx) + "/" + _PROJECT_REPO_DIR
 
     # For each config repo we run the target @<config_repo>//...
     bazel_targets = "@" + "//... @".join(config_repos) + "//..."
-    bazel_cmd += " && " + _BAZELISK_PATH + " build " + bazel_targets
+
+    # TODO: the target exclusion below is b/c bazel has a bug, see: https://github.com/bazelbuild/bazel/issues/11101
+    # the relevant changes have not made their way to a bazel release yet
+    if _is_windows(ctx):
+        bazel_targets += " -- -@local_config_cc//:link_dynamic_library"
+
+    bazel_cmd += " && " + _bazelisk_path(ctx) + " build " + bazel_targets
 
     # Command to run to clean up after autoconfiguration.
     # we start with "cd ." to make sure in case of failure everything after the
     # ";" will be executed
     clean_cmd = "cd . ; bazel clean"
-    if use_default_project:
-        clean_cmd += "; rm WORKSPACE ; mv BUILD BUILD.sample"
 
     docker_cmd = [
         "#!/bin/bash",
@@ -172,7 +202,7 @@ def get_java_home(ctx, docker_tool_path, image_name):
         )
 
         # run get_java_home.sh
-        result = ctx.execute(["./get_java_home.sh"])
+        result = ctx.execute(["bash", "./get_java_home.sh"])
         print_exec_results("get java_home", result, fail_on_error = True)
         java_home = result.stdout.splitlines()[0]
         if java_home == "":
@@ -185,6 +215,40 @@ def get_java_home(ctx, docker_tool_path, image_name):
               "detect_java_home was set.") % ctx.attr.name)
     else:
         return None
+
+def _generate_sample_cc_project(ctx):
+    ctx.file(
+        "cc-sample-project/BUILD",
+        """package(default_visibility = ["//visibility:public"])
+
+licenses(["notice"])  # Apache 2.0
+
+filegroup(
+    name = "srcs",
+    srcs = [
+        "BUILD",
+        "test.cc",
+    ],
+)
+
+cc_test(
+    name = "test",
+    srcs = ["test.cc"],
+)
+""",
+    )
+    ctx.file(
+        "cc-sample-project/test.cc",
+        """#include <iostream>
+
+int main() {
+  std::cout << "Hello test!" << std::endl;
+  return 0;
+}
+
+""",
+    )
+    ctx.file("cc-sample-project/WORKSPACE", "")
 
 def run_and_extract(
         ctx,
@@ -224,9 +288,9 @@ def run_and_extract(
         use_default_project = use_default_project,
     )
 
-    # Download bazelisk
-    bazelisk_url = "https://github.com/bazelbuild/bazelisk/releases/download/%s/bazelisk-linux-amd64" % _BAZELISK_RELEASE
-    ctx.download(bazelisk_url, "bazelisk", _BAZELISK_SHA)
+    _download_bazelisk(ctx)
+
+    _generate_sample_cc_project(ctx)
 
     # Create the docker run flags to set env vars
     docker_run_flags = []
@@ -239,17 +303,32 @@ def run_and_extract(
     # Set the Bazel version that Bazelisk will use
     docker_run_flags += ["--env", ("USE_BAZEL_VERSION=%s" % bazel_version_string)]
 
+    # Run as the root user on Linux and equivalent ContainerAdministrator on Windows for images
+    # that have a default USER set in the Dockerfile
+    if _is_windows(ctx):
+        docker_run_flags += ["--user", "ContainerAdministrator"]
+    else:
+        docker_run_flags += ["--user", "root"]
+
     # Command to copy resources used for rbe_autoconfig to the container.
     copy_data_cmd = []
 
     # Command to clean up the data volume container.
     clean_data_volume_cmd = ""
-    copy_data_cmd.append("data_volume=$(docker create -v " + _ROOT_DIR + " " + image_name + ")")
-    copy_data_cmd.append(docker_tool_path + " cp $(realpath " + project_root + ") $data_volume:" + _REPO_DIR)
-    copy_data_cmd.append(docker_tool_path + " cp " + str(ctx.path("container")) + " $data_volume:" + _ROOT_DIR + "/container")
-    copy_data_cmd.append(docker_tool_path + " cp " + str(ctx.path("bazelisk")) + " $data_volume:" + _BAZELISK_PATH)
-    docker_run_flags += ["--volumes-from", "$data_volume"]
-    clean_data_volume_cmd = docker_tool_path + " rm $data_volume"
+
+    # docker cp does not function on Windows as expected when copying into volumes so we use bind
+    # mounts instead
+    if _is_windows(ctx):
+        docker_run_flags += ["-v", str(ctx.path("cc-sample-project")) + ":" + _root_dir(ctx) + "/" + _PROJECT_REPO_DIR]
+        docker_run_flags += ["-v", str(ctx.path("container")) + ":" + _root_dir(ctx) + "/container"]
+        docker_run_flags += ["-v", str(ctx.path("bazelisk")) + ":" + _root_dir(ctx) + "/bazelisk"]
+    else:
+        copy_data_cmd.append("data_volume=$(docker -v " + _root_dir(ctx) + " " + image_name + ")")
+        copy_data_cmd.append(docker_tool_path + " cp " + str(ctx.path("cc-sample-project")) + " $data_volume:" + _root_dir(ctx) + "/" + _PROJECT_REPO_DIR)
+        copy_data_cmd.append(docker_tool_path + " cp " + str(ctx.path("container")) + " $data_volume:" + _root_dir(ctx) + "/container")
+        copy_data_cmd.append(docker_tool_path + " cp " + str(ctx.path(_BAZELISK_FILENAME_LINUX)) + " $data_volume:" + _bazelisk_path(ctx))
+        docker_run_flags += ["--volumes-from", "$data_volume"]
+        clean_data_volume_cmd = docker_tool_path + " rm $data_volume"
 
     # Create the template to run
     template = ctx.path(Label("@bazel_toolchains//rules/rbe_repo:extract.sh.tpl"))
@@ -258,7 +337,7 @@ def run_and_extract(
         template,
         {
             "%{clean_data_volume_cmd}": clean_data_volume_cmd,
-            "%{commands}": _ROOT_DIR + "/container/run_in_container.sh",
+            "%{commands}": "bash " + _root_dir(ctx) + "/container/run_in_container.sh",
             "%{copy_data_cmd}": " && ".join(copy_data_cmd),
             "%{docker_run_flags}": " ".join(docker_run_flags),
             "%{docker_tool_path}": docker_tool_path,
@@ -271,7 +350,7 @@ def run_and_extract(
 
     # run run_and_extract.sh
     ctx.report_progress("running container")
-    result = ctx.execute(["./run_and_extract.sh"])
+    result = ctx.execute(["bash", "./run_and_extract.sh"])
     print_exec_results("run_and_extract", result, fail_on_error = True)
 
     # Expand outputs inside this remote repo
